requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-filter==2.2.0
django-mdeditor==0.1.16
django-registration==3.1
djangorestframework==3.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==7.0.0
Pygments==2.5.2
pytz==2019.3
PyYAML==5.3
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
